# backend/app/admin/service/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: GoWind Admin API
    description: GoWind Admin API
    contact:
        name: tx7do
        url: https://github.com/tx7do/go-wind-admin
        email: yanglinbo@gmail.com
    license:
        name: MIT License
        url: https://github.com/tx7do/go-wind-admin/blob/master/LICENSE
    version: "1.0"
paths:
...
    /admin/v1/tenants:
        get:
            tags:
                - TenantService
            description: 获取租户列表
            operationId: TenantService_List
            parameters:
                - name: page
                  in: query
                  description: 当前页码（从1开始，默认1）
                  schema:
                    type: integer
                    format: uint32
                - name: pageSize
                  in: query
                  description: 每页条数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: offset
                  in: query
                  description: 跳过的记录数（从0开始，默认0）
                  schema:
                    type: string
                - name: limit
                  in: query
                  description: 最多返回的记录数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: token
                  in: query
                  description: 上一页最后一条记录的游标（如ID/时间戳+ID，首次请求为空）
                  schema:
                    type: string
                - name: noPaging
                  in: query
                  description: 是否不分页，如果为true，则page和pageSize参数无效。
                  schema:
                    type: boolean
                - name: query
                  in: query
                  description: JSON字符串过滤条件，基础语法：{"field1":"val1", "field2___icontains":"val2"}，具体请参见：https://github.com/tx7do/go-crud/tree/main/pagination/filter/README.md
                  schema:
                    type: string
                - name: filter
                  in: query
                  description: Google AIP规范字符串过滤条件
                  schema:
                    type: string
                - name: filterExpr.type
                  in: query
                  description: 过滤表达式类型
                  schema:
                    enum:
                        - EXPR_TYPE_UNSPECIFIED
                        - AND
                        - OR
                    type: string
                    format: enum
                - name: orderBy
                  in: query
                  description: 排序条件
                  schema:
                    type: string
                - name: fieldMask
                  in: query
                  description: 字段掩码，其作用为SELECT中的字段，其语法为使用逗号分隔字段名，例如：id,realName,userName。如果为空则选中所有字段，即SELECT *。
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListTenantResponse'
        post:
            tags:
                - TenantService
            description: 创建租户
            operationId: TenantService_Create
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateTenantRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /admin/v1/tenants/{id}:
        get:
            tags:
                - TenantService
            description: 获取租户数据
            operationId: TenantService_Get
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: code
                  in: query
                  schema:
                    type: string
                - name: name
                  in: query
                  schema:
                    type: string
                - name: viewMask
                  in: query
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Tenant'
        put:
            tags:
                - TenantService
            description: 更新租户
            operationId: TenantService_Update
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateTenantRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
        delete:
            tags:
                - TenantService
            description: 删除租户
            operationId: TenantService_Delete
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content: {}
    /admin/v1/tenants:exists:
        get:
            tags:
                - TenantService
            description: 租户是否存在
            operationId: TenantService_TenantExists
            parameters:
                - name: code
                  in: query
                  schema:
                    type: string
                - name: name
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TenantExistsResponse'
    /admin/v1/tenants:with-admin:
        post:
            tags:
                - TenantService
            description: 创建租户及管理员用户
            operationId: TenantService_CreateTenantWithAdminUser
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateTenantWithAdminUserRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
...
components:
    schemas:
...
        CreateTenantRequest:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/Tenant'
            description: 创建租户 - 请求
        CreateTenantWithAdminUserRequest:
            type: object
            properties:
                tenant:
                    $ref: '#/components/schemas/Tenant'
                user:
                    $ref: '#/components/schemas/User'
                password:
                    type: string
            description: 创建租户及管理员用户 - 请求
...
        ListTenantResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/Tenant'
                total:
                    type: string
            description: 租户列表 - 答复
...
        Tenant:
            type: object
            properties:
                id:
                    type: integer
                    description: 租户ID
                    format: uint32
                name:
                    type: string
                    description: 租户名称
                code:
                    type: string
                    description: 租户编码
                domain:
                    type: string
                    description: 租户专属域名
                logoUrl:
                    type: string
                    description: 租户logo地址
                industry:
                    type: string
                    description: 所属行业（如“互联网”“金融”“制造”）
                type:
                    enum:
                        - TENANT_TYPE_UNSPECIFIED
                        - TRIAL
                        - PAID
                        - INTERNAL
                        - PARTNER
                        - CUSTOM
                    type: string
                    description: 租户类型
                    format: enum
                remark:
                    type: string
                    description: 备注
                adminUserId:
                    type: integer
                    description: 管理员用户ID
                    format: uint32
                adminUserName:
                    type: string
                    description: 管理员用户名
                subscriptionAt:
                    type: string
                    description: 订阅时间（首次订阅/续费时间，NULL表示未订阅）
                    format: date-time
                unsubscribeAt:
                    type: string
                    description: 取消订阅时间
                    format: date-time
                expiredAt:
                    type: string
                    description: 租户有效期（NULL表示永久，过期后状态自动改为“过期”）
                    format: date-time
                subscriptionPlan:
                    type: string
                    description: 订阅套餐（如“企业版1年”“基础版3个月”）
                memberCount:
                    type: integer
                    description: 成员数量
                    format: int32
                status:
                    enum:
                        - OFF
                        - ON
                        - EXPIRED
                        - FREEZE
                    type: string
                    description: 租户状态
                    format: enum
                auditStatus:
                    enum:
                        - TENANT_AUDIT_STATUS_UNSPECIFIED
                        - PENDING
                        - APPROVED
                        - REJECTED
                    type: string
                    description: 审核状态
                    format: enum
                createdBy:
                    type: integer
                    description: 创建者ID
                    format: uint32
                updatedBy:
                    type: integer
                    description: 更新者ID
                    format: uint32
                deletedBy:
                    type: integer
                    description: 删除者用户ID
                    format: uint32
                createdAt:
                    type: string
                    description: 创建时间
                    format: date-time
                updatedAt:
                    type: string
                    description: 更新时间
                    format: date-time
                deletedAt:
                    type: string
                    description: 删除时间
                    format: date-time
            description: 租户
        TenantExistsResponse:
            type: object
            properties:
                exist:
                    type: boolean
            description: 租户是否存在 - 答复
...
        UpdateTenantRequest:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                data:
                    $ref: '#/components/schemas/Tenant'
                updateMask:
                    example: id,realname,username
                    type: string
                    description: 要更新的字段列表
                    format: field-mask
                allowMissing:
                    type: boolean
                    description: 如果设置为true的时候，资源不存在则会新增(插入)，并且在这种情况下`updateMask`字段将会被忽略。
            description: 更新租户 -请求
...
        User:
            type: object
            properties:
                id:
                    type: integer
                    description: 用户ID
                    format: uint32
                tenantId:
                    type: integer
                    description: 租户ID
                    format: uint32
                tenantName:
                    type: string
                    description: 租户名称
                orgUnitId:
                    type: integer
                    description: 组织ID
                    format: uint32
                orgUnitIds:
                    type: array
                    items:
                        type: integer
                        format: uint32
                    description: 归属组织ID列表
                orgUnitName:
                    type: string
                    description: 组织名称
                orgUnitNames:
                    type: array
                    items:
                        type: string
                    description: 组织名称列表
                positionId:
                    type: integer
                    description: 职位ID
                    format: uint32
                positionIds:
                    type: array
                    items:
                        type: integer
                        format: uint32
                    description: 职位列表
                positionName:
                    type: string
                    description: 职位名称
                positionNames:
                    type: array
                    items:
                        type: string
                    description: 职位名称列表
                roleId:
                    type: integer
                    description: 角色ID
                    format: uint32
                roleIds:
                    type: array
                    items:
                        type: integer
                        format: uint32
                    description: 角色ID列表
                roles:
                    type: array
                    items:
                        type: string
                    description: 角色码列表
                roleNames:
                    type: array
                    items:
                        type: string
                    description: 角色名称列表
                username:
                    type: string
                    description: 用户名
                nickname:
                    type: string
                    description: 昵称
                realname:
                    type: string
                    description: 真实姓名
                avatar:
                    type: string
                    description: 头像
                email:
                    type: string
                    description: 邮箱
                mobile:
                    type: string
                    description: 手机号
                telephone:
                    type: string
                    description: 座机号
                gender:
                    enum:
                        - SECRET
                        - MALE
                        - FEMALE
                    type: string
                    description: 性别
                    format: enum
                address:
                    type: string
                    description: 住址
                region:
                    type: string
                    description: 国家地区
                description:
                    type: string
                    description: 个人描述
                remark:
                    type: string
                    description: 备注
                lastLoginAt:
                    type: string
                    description: 最后登录时间
                    format: date-time
                lastLoginIp:
                    type: string
                    description: 最后登录IP
                status:
                    enum:
                        - DISABLED
                        - NORMAL
                        - PENDING
                        - LOCKED
                        - EXPIRED
                        - CLOSED
                    type: string
                    description: 状态
                    format: enum
                lockedUntil:
                    type: string
                    description: 锁定截止时间
                    format: date-time
                createdBy:
                    type: integer
                    description: 创建者ID
                    format: uint32
                updatedBy:
                    type: integer
                    description: 更新者ID
                    format: uint32
                deletedBy:
                    type: integer
                    description: 删除者用户ID
                    format: uint32
                createdAt:
                    type: string
                    description: 创建时间
                    format: date-time
                updatedAt:
                    type: string
                    description: 更新时间
                    format: date-time
                deletedAt:
                    type: string
                    description: 删除时间
                    format: date-time
            description: 用户
...
    securitySchemes:
        OAuth2PasswordBearer:
            type: oauth2
            flows:
                password:
                    tokenUrl: /admin/v1/login
                    refreshUrl: /admin/v1/refresh-token
                    scopes: {}
security:
    - OAuth2PasswordBearer: []
...